+++ platform/android/dnsproxy/lib/src/main/cpp/android_dnsproxy.cpp
@@ -97,9 +97,10 @@ Java_com_adguard_dnslibs_proxy_DnsProxy_isValidRule(JNIEnv *env, jclass clazz, j
 extern "C"
 JNIEXPORT jstring JNICALL
 Java_com_adguard_dnslibs_proxy_DnsProxy_testUpstreamNative(JNIEnv *env, jclass clazz, jlong native_ptr,
-                                                           jobject upstream_settings, jobject events_adapter) {
+                                                           jobject upstream_settings, jboolean ipv6,
+                                                           jobject events_adapter) {
     auto *proxy = (ag::android_dnsproxy *) native_ptr;
-    return proxy->test_upstream(env, upstream_settings, events_adapter);
+    return proxy->test_upstream(env, upstream_settings, ipv6, events_adapter);
 }
 
 extern "C"
@@ -769,9 +770,9 @@ jobject ag::android_dnsproxy::get_settings(JNIEnv *env) {
     return env->NewLocalRef(marshal_settings(env, m_actual_proxy.get_settings()).get());
 }
 
-jstring ag::android_dnsproxy::test_upstream(JNIEnv *env, jobject upstream_settings, jobject events_adapter) {
+jstring ag::android_dnsproxy::test_upstream(JNIEnv *env, jobject upstream_settings, jboolean ipv6, jobject events_adapter) {
     m_events = global_ref(get_vm(env), events_adapter);
-    auto err = ag::test_upstream(marshal_upstream(env, upstream_settings),
+    auto err = ag::test_upstream(marshal_upstream(env, upstream_settings), ipv6,
                                  marshal_events(env, events_adapter).on_certificate_verification);
     if (err) {
         return (jstring) env->NewLocalRef(m_utils.marshal_string(env, *err).get());