+++ platform/android/dnsproxy/lib/src/main/cpp/android_dnsproxy.h
@@ -182,9 +182,9 @@ public:
 
     /**
      * Checks if upstream is valid and available.
-     * @return Null or error string marshaleed to Java.
+     * @return Null or error string marshalled to Java.
      */
-    jstring test_upstream(JNIEnv *env, jobject upstream_settings, jobject events_adapter);
+    jstring test_upstream(JNIEnv *env, jobject upstream_settings, jboolean ipv6, jobject events_adapter);
 };
 
 }