+++ platform/mac/framework/AGDnsProxy.mm
@@ -1409,8 +1409,9 @@ static std::optional<std::string> verifyCertificate(ag::certificate_verification
 }
 
 + (NSError *) testUpstream: (AGDnsUpstream *) opts
+             ipv6Available: (BOOL) ipv6Available
 {
-    auto error = ag::test_upstream(convert_upstream(opts), verifyCertificate);
+    auto error = ag::test_upstream(convert_upstream(opts), ipv6Available, verifyCertificate);
     if (error) {
         return [NSError errorWithDomain: AGDnsProxyErrorDomain
                                    code: AGDPE_TEST_UPSTREAM_ERROR