+++ platform/windows/capi/test/capi_test.c
@@ -151,13 +151,13 @@ static void test_utils() {
     upstream.bootstrap.data = malloc(sizeof(const char *));
     upstream.bootstrap.data[0] = "8.8.8.8";
     upstream.timeout_ms = 5000;
-    const char *error = ag_test_upstream(&upstream, on_cert);
+    const char *error = ag_test_upstream(&upstream, false, on_cert);
     ASSERT(error == NULL);
     upstream.address = "1.2.3.4.5.6";
-    error = ag_test_upstream(&upstream, NULL);
+    error = ag_test_upstream(&upstream, false, NULL);
     ASSERT(error);
     upstream.address = "https://asdf.asdf.asdf/asdfdnsqueryasdf";
-    error = ag_test_upstream(&upstream, NULL);
+    error = ag_test_upstream(&upstream, false, NULL);
     ASSERT(error);
     ag_str_free(error);
     free(upstream.bootstrap.data);