+++ upstream/include/upstream_utils.h
@@ -11,17 +11,11 @@ namespace ag {
 /**
  * Checks if upstream is valid and available
  * @param opts Upstream options
+ * @param ipv6_available whether IPv6 is available, i.e. bootstrapper should make AAAA queries
  * @param on_certificate_verification Certificate verification callback
  * @return If it is, no error is returned. Otherwise this method returns an error with an explanation.
  */
-err_string test_upstream(const upstream_options &opts,
+err_string test_upstream(const upstream_options &opts, bool ipv6_available,
                          const on_certificate_verification_function &on_certificate_verification);
-
-/**
- * Test if a well-known plain DNS server is reachable over IPv6.
- * @return true if IPv6 works,
- *         false otherwise.
- */
-bool test_ipv6_connectivity();
 
 } // namespace ag